% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{The boot2pval function. Creates a bootstrap interval for a two sample test. Additionally outputs a graph showing the p-value.}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{the first vector sample.}

\item{x2}{the second vector sample.}

\item{conf.level}{the confdience level for the confidence interval.}

\item{iter}{the number of iterations done in the bootstrap testing.}

\item{mudiff}{the tested/expected difference in means between the two samples.}

\item{test}{the type of test e.g two tailed or one tailed.}
}
\value{
An output giving the pvalue as well as the graph displaying the t distribution.
}
\description{
The boot2pval function. Creates a bootstrap interval for a two sample test. Additionally outputs a graph showing the p-value.
}
\examples{
boot2pval(x1=x, x2=y, mudiff=0)
}
