% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{The Mode Function}
\usage{
getmode(x)
}
\arguments{
\item{x}{a vector of quantitative data}
}
\value{
the mode of x
}
\description{
The Mode Function
}
\examples{
getmode(c(1, 3, 4, 4))
}
