% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{The mybin function. Creates a binomial simulation with replacement.}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{an integer number of iterations for each graph}

\item{n}{an integer number of the amount of successes that will be tested.}

\item{p}{an integer value from 0 <= p <= 1 that defines the probability of success with each trial.}
}
\value{
Binomial graphs created with the constraints provided.
}
\description{
The mybin function. Creates a binomial simulation with replacement.
}
\examples{
mybin(iter=100, n=10, p=0.5)
}
