% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{The myboot2 function. Creates a bootstrap simulation to calculate the confidence interval of unknown distributions.}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5)
}
\arguments{
\item{iter}{an integer number of iterations that are run in the simulation}

\item{x}{the distribution being tested.}

\item{fun}{the function which is used in the calculations.}

\item{alpha}{the alpha value used in the (1-alpha)100% confidence interval}

\item{cx}{used as a density statistic.}
}
\value{
A graph showing the estimated confidence interval of the boostrap simulation.
}
\description{
The myboot2 function. Creates a bootstrap simulation to calculate the confidence interval of unknown distributions.
}
\examples{
myboot2(iter=10000,x=sam,fun=function(x) mean(x)/median(x))
}
