% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{The myci function. Creates a confidence interval of the sample mean.}
\usage{
myci(d, confInt)
}
\arguments{
\item{d}{a vector sample.}

\item{confInt}{the confidence interval. from 0 - 1.}
}
\value{
A confidence interval of the sample mean returned as a vector upper, lower
}
\description{
The myci function. Creates a confidence interval of the sample mean.
}
\examples{
myci(d, 0.95)
}
