% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcltp.R
\name{mycltp}
\alias{mycltp}
\title{The MyCltp Function. Creates a CLT graph of a Poisson distribution.}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{an integer stating the number of trails in the sample.}

\item{iter}{an integer that sets the amount of different iterations that will be run.}

\item{lambda}{the lambda value of the poisson distribution.}
}
\value{
a graphical output showing the sampled distributions.
}
\description{
The MyCltp Function. Creates a CLT graph of a Poisson distribution.
}
\examples{
mycltp(n=4, iter=10000, lambda=10);
}
