% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{The mymlnorm function. Makesa max lik function for two parameters}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{the sample getting tested}

\item{mu}{the mean of the sample}

\item{sig}{the standard deviation of the sample}
}
\value{
A graph showing the estimated maximum likliehood.
}
\description{
The mymlnorm function. Makesa max lik function for two parameters
}
\examples{
mymlnorm(x=c(5,7,7,8,10),mu=seq(5,10,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)
}
