% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{The myncurve Function}
\usage{
myncurve(mu, sigma, a, listnum = 1)
}
\arguments{
\item{mu}{the mean of the data}

\item{sigma}{the standard deviation of the data}

\item{a}{the upper bound of the area calculation}

\item{listnum}{the number 1<=listnum<=3 that selects the output from the linked list}
}
\value{
a graphical representation of the curve along with a named list output.
}
\description{
The myncurve Function
}
\examples{
myncurve(10, 3, 1, 2)
}
